Split JavaScript runner
