+++ lib/runners/javascript/index.js
@@ -5,7 +5,12 @@ const path = require('path');
 const fs = require('fs-extra');
 
 const writeFilesSync = require('../../utils/write-files-sync');
-const frontendFrameworks = require('../../frontend-frameworks');
+const frontendFrameworks = require('./frontend-frameworks');
+const transform = require('./transform');
+const maybeTransform = require('./maybe-transform');
+const createNodeRunCommand = require('./create-node-run-command');
+const execNode = require('./exec-node');
+const runKarma = require('./run-karma');
 
 module.exports = {
   solutionOnly(opts, runCode, fail) {
@@ -129,44 +134,8 @@ function prepareKarma(opts, interfaceType, runCode, fail) {
   });
 }
 
-function runKarma(opts, runCode, fail, interfaceType, config) {
-  config = Object.assign({
-    singleRun: true,
-    autoWatch: false,
-    retryLimit: 0,
-    frameworks: ['mocha', 'chai'],
-    browsers: ['PhantomJS'],
-    plugins: ['karma-*', '/runner/frameworks/javascript/karma-coderunner-reporter.js'],
-    reporters: ['coderunner'],
-    logLevel: 'warn',
-    client: {
-      mocha: {
-        timeout: opts.timeout || 10000,
-        ui: interfaceType,
-      },
-    },
-  }, config);
-
-  let sConfig = JSON.stringify(config, null, 4)
-    // convert any strings that start with `@@` to literal JS
-    .replace(/"@@([^"]+)"/g, (_, literal) => literal.replace(/\\(.)/g, '$1'));
-
-  var code = `
-        var KarmaServer = require('karma').Server;
-        var karma = new KarmaServer(${sConfig}, function(code) {
-            process.exit(code);
-        });
-        karma.start();
-    `;
-
-  execNode(opts, code, runCode, fail);
-}
-// TODO separate this
-module.exports.runKarma = runKarma;
-
 function prepareCw2(opts, runCode, fail) {
   var code;
-
   if (opts.projectMode) {
     code = `
             require('/runner/frameworks/javascript/cw-2');
@@ -195,140 +164,4 @@ function prepareCw2(opts, runCode, fail) {
   }
 
   execNode(opts, code, runCode, fail);
-}
-
-function execNode(opts, code, runCode, fail) {
-  try {
-    code = maybeTransform(code, opts);
-
-    var args = ['-e', code, '--no-deprecation'];
-    // if (version.name == '6.x') {
-    //     args.push('--trace-warnings');
-    // }
-
-    (opts.requires || []).forEach(function(f) {
-      args.push("--require");
-      args.push(f);
-    });
-
-    runCode(createNodeRunCommand(args, opts));
-  }
-  catch (ex) {
-    fail(ex);
-  }
-}
-
-function createNodeRunCommand(args, opts) {
-  var version = versionInfo(opts);
-  var nodeCmd = `/usr/local/n/versions/node/${version.node}/bin/node`;
-  if (!fs.existsSync(nodeCmd)) {
-    throw new Error('Invalid node version: ' + version.node + '\n');
-  }
-  return {
-    name: nodeCmd,
-    args: args,
-    options: {
-      cwd: opts.dir
-    }
-  };
-}
-
-function nodeVersion(version) {
-  switch (version) {
-    case "8.x":
-      return "8.1.3";
-    case "6.x":
-    case "6.6.0": // legacy version mapping
-      return "6.11.0";
-    case "0.10.x":
-      return "0.10.33";
-    default:
-      return version;
-  }
-}
-
-// returns information about the version passed in, such as its node version and if babel is enabled
-function versionInfo(opts) {
-  var info = {
-    name: (opts.languageVersion || '6.x').split('/')[0],
-    babel: opts.languageVersion && opts.languageVersion.split('/')[1] == 'babel'
-  };
-  info.node = nodeVersion(info.name);
-  return info;
-}
-
-// will babel transform the code if babel is configured, otherwise will just return the code unaltered.
-function maybeTransform(code, opts, filename) {
-  var version = versionInfo(opts);
-  if (version.babel) {
-    return transform(code, version.name, filename);
-  }
-  return code;
-}
-
-function transform(code, version, filename) {
-  try {
-    switch (version) {
-      case '0.10.x':
-        return require('babel-core').transform(code, {
-          presets: ["stage-1", "react"],
-          plugins: [
-            "check-es2015-constants",
-            "angular2-annotations",
-            "transform-decorators-legacy",
-            "transform-class-properties",
-            "transform-flow-strip-types",
-            "transform-es2015-arrow-functions",
-            "transform-es2015-block-scoped-functions",
-            "transform-es2015-block-scoping",
-            "transform-es2015-classes",
-            "transform-es2015-computed-properties",
-            "transform-es2015-destructuring",
-            "transform-es2015-duplicate-keys",
-            "transform-es2015-for-of",
-            "transform-es2015-function-name",
-            "transform-es2015-literals",
-            "transform-es2015-object-super",
-            "transform-es2015-parameters",
-            "transform-es2015-shorthand-properties",
-            "transform-es2015-spread",
-            "transform-es2015-sticky-regex",
-            "transform-es2015-template-literals",
-            "transform-es2015-typeof-symbol",
-            "transform-es2015-unicode-regex",
-            "transform-regenerator",
-          ],
-          ast: false,
-          filename: filename || 'kata.js'
-        }).code;
-
-      default:
-        return require('babel-core').transform(code, {
-          presets: ["stage-1", "node5", "react"],
-          plugins: [
-            "angular2-annotations",
-            "transform-decorators-legacy",
-            "transform-class-properties",
-            "transform-flow-strip-types",
-          ],
-          ast: false,
-          filename: filename || 'kata.js'
-        }).code;
-    }
-  }
-  catch (ex) {
-    var msg = ex.message;
-    // if (ex.loc) {
-    //     // replace the line number since it is not what the user sees
-    //     msg = msg.replace(/ \(\d*:\d*\)/, ":" + ex.loc.column)
-    //     var lines = code.split('\n');
-    //     msg += "\n" + lines[ex.loc.line - 1];
-    //     msg += "\n";
-    //     for(var i = 1;i < ex.loc.column; i++) {
-    //         msg += ' ';
-    //     }
-    //     msg += '^';
-    // }
-    throw new Error(msg);
-  }
 }
