+++ lib/runners/typescript.js
@@ -5,8 +5,8 @@ const exec = require('child_process').exec;
 
 const fs = require('fs-extra');
 
-const frontendFrameworks = require('../frontend-frameworks');
-const runKarma = require('./javascript').runKarma;
+const frontendFrameworks = require('./javascript/frontend-frameworks');
+const runKarma = require('./javascript/run-karma');
 
 module.exports = {
   solutionOnly(opts, runCode, fail) {
